header

+++ ft_putnbr_fd.c
@@ -1,7 +1,7 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   putnbr_fd.c                                        :+:      :+:    :+:   */
+/*   ft_putnbr_fd.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: idouidi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
@@ -14,13 +14,13 @@
 
 void	ft_putnbr_fd(int n, int fd)
 {
-	unsigned int res;
+	long	res;
 
-	res = (unsigned int)n;
+	res = (long)n;
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
-		res = (unsigned int)(n * -1);
+		res = (long)(n * -1);
 	}
 	if (res >= 0 && res <= 9)
 		ft_putchar_fd(res + '0', fd);
